models/deepF-img-new/requirements.txt
fastapi 
uvicorn 
torch 
torchvision 
transformers 
pillow 
opencv-python
python-multipart